requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.23
pyyaml>=6.0
pillow>=9.0
tqdm>=4.65